Add support for writing LAZ with variable-size chunks
- LasZipCompressor supports writting LAZ files with variable chunk size
- ParLasZipCompressor supports writting LAZ files with variable chunk size
- Regorganize `laszip` module
