+++ src/las/mod.rs
@@ -2,9 +2,11 @@
 //! LAS data compressors & decompressors as well as
 //! the definition of the point types.
 
+pub use pointtypes::{Point0, Point1, Point2, Point3, Point6, Point7, Point8};
+
 #[macro_use]
 mod utils;
-mod pointtypes;
+pub(crate) mod pointtypes;
 
 pub mod point0;
 pub mod point6;
@@ -12,11 +14,8 @@ pub mod point6;
 pub mod extra_bytes;
 pub mod file;
 pub mod gps;
-pub mod laszip;
 pub mod nir;
 pub mod rgb;
-
-pub use pointtypes::{Point0, Point1, Point2, Point3, Point6, Point7, Point8};
 
 pub mod v1 {
     //! This module only contains re exports of compressors / decompressors
