+++ src/record.rs
@@ -8,7 +8,7 @@ use crate::byteslice::{ChunksIrregular, ChunksIrregularMut};
 use crate::decoders;
 use crate::encoders;
 use crate::las;
-use crate::las::laszip::{LazItem, LazItemType};
+use crate::laszip::{LazItem, LazItemType};
 use crate::LasZipError;
 
 /***************************************************************************************************
